Changed labels to standard
Diplomacy was originally published in the United States in 1959. As such any features of the original map, including the unique combination of city and region names for game spaces and the exaggerated importance of Turkish provinces for the sake of game balance, that might have had enforceable copyright are no longer afforded copyright protection.